improves measure docs
